backend/requirements.txt
fastapi
uvicorn
python-multipart
pycryptodome
numpy
pynput
